requirements.txt
boto3
pyspark
pandas
psycopg2-binary
python-dotenv
pyarrow
apache-airflow